requirements.txt
requests
PyYAML
flask
docker